+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 ### Version 9.5
 - Show ingredients label text in modal in finished product section
+- Move batch history menu under formulas
+- Fix typos in IFRA generation document
 
 ### Version 9.4
 - Fix update message when in cloud enviroment